+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.4.10 (Unreleased)
 - [Enhancement] Provide `Karafka::Messages::Message#tombstone?` to be able to quickly check if a message is a tombstone message.
+- [Enhancement] Provide more flexible API for Recurring Tasks topics reconfiguration.
 
 ## 2.4.9 (2024-08-23)
 - **[Feature]** Provide Kafka based Recurring (Cron) Tasks.